# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API documentation",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/pdf/sync/html": {
      "post": {
        "operationId": "generatePdfByHTML",
        "tags": [
          "PDF"
        ],
        "description": "Generate PDF from HTML",
        "parameters": [
          {
            "schema": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "scale": {
                  "type": "number",
                  "minimum": 0.1,
                  "maximum": 2
                },
                "displayHeaderFooter": {
                  "type": "boolean"
                },
                "headerTemplate": {
                  "type": "string"
                },
                "footerTemplate": {
                  "type": "string"
                },
                "printBackground": {
                  "type": "boolean"
                },
                "landscape": {
                  "type": "boolean"
                },
                "pageRanges": {
                  "type": "string"
                },
                "format": {
                  "type": "string",
                  "enum": [
                    "Letter",
                    "Legal",
                    "Tabloid",
                    "Ledger",
                    "A0",
                    "A1",
                    "A2",
                    "A3",
                    "A4",
                    "A5",
                    "A6"
                  ]
                },
                "width": {
                  "type": "string",
                  "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                },
                "height": {
                  "type": "string",
                  "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                },
                "margin": {
                  "type": "object",
                  "properties": {
                    "top": {
                      "type": "string"
                    },
                    "right": {
                      "type": "string"
                    },
                    "bottom": {
                      "type": "string"
                    },
                    "left": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                },
                "preferCSSPageSize": {
                  "type": "boolean"
                },
                "omitBackground": {
                  "type": "boolean"
                }
              },
              "additionalProperties": false
            },
            "in": "query",
            "name": "options",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/pdf/sync/url": {
      "post": {
        "operationId": "generatePdfByUrl",
        "tags": [
          "PDF"
        ],
        "description": "Generate PDF from URL",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "string"
                      },
                      "scale": {
                        "type": "number",
                        "minimum": 0.1,
                        "maximum": 2
                      },
                      "displayHeaderFooter": {
                        "type": "boolean"
                      },
                      "headerTemplate": {
                        "type": "string"
                      },
                      "footerTemplate": {
                        "type": "string"
                      },
                      "printBackground": {
                        "type": "boolean"
                      },
                      "landscape": {
                        "type": "boolean"
                      },
                      "pageRanges": {
                        "type": "string"
                      },
                      "format": {
                        "type": "string",
                        "enum": [
                          "Letter",
                          "Legal",
                          "Tabloid",
                          "Ledger",
                          "A0",
                          "A1",
                          "A2",
                          "A3",
                          "A4",
                          "A5",
                          "A6"
                        ]
                      },
                      "width": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                      },
                      "height": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                      },
                      "margin": {
                        "type": "object",
                        "properties": {
                          "top": {
                            "type": "string"
                          },
                          "right": {
                            "type": "string"
                          },
                          "bottom": {
                            "type": "string"
                          },
                          "left": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      },
                      "preferCSSPageSize": {
                        "type": "boolean"
                      },
                      "omitBackground": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": [
                  "url"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/pdf/async/html": {
      "post": {
        "operationId": "generatePdfByHTMLAsync",
        "tags": [
          "PDF"
        ],
        "description": "Generate PDF from HTML asynchronously",
        "parameters": [
          {
            "schema": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "scale": {
                  "type": "number",
                  "minimum": 0.1,
                  "maximum": 2
                },
                "displayHeaderFooter": {
                  "type": "boolean"
                },
                "headerTemplate": {
                  "type": "string"
                },
                "footerTemplate": {
                  "type": "string"
                },
                "printBackground": {
                  "type": "boolean"
                },
                "landscape": {
                  "type": "boolean"
                },
                "pageRanges": {
                  "type": "string"
                },
                "format": {
                  "type": "string",
                  "enum": [
                    "Letter",
                    "Legal",
                    "Tabloid",
                    "Ledger",
                    "A0",
                    "A1",
                    "A2",
                    "A3",
                    "A4",
                    "A5",
                    "A6"
                  ]
                },
                "width": {
                  "type": "string",
                  "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                },
                "height": {
                  "type": "string",
                  "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                },
                "margin": {
                  "type": "object",
                  "properties": {
                    "top": {
                      "type": "string"
                    },
                    "right": {
                      "type": "string"
                    },
                    "bottom": {
                      "type": "string"
                    },
                    "left": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                },
                "preferCSSPageSize": {
                  "type": "boolean"
                },
                "omitBackground": {
                  "type": "boolean"
                }
              },
              "additionalProperties": false
            },
            "in": "query",
            "name": "options",
            "required": false
          }
        ],
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "jobIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "message",
                    "jobIds",
                    "errors"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/pdf/async/url": {
      "post": {
        "operationId": "generatePdfByUrlAsync",
        "tags": [
          "PDF"
        ],
        "description": "Generate PDF from URL asynchronously",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "urls": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string",
                          "format": "uri"
                        }
                      },
                      "required": [
                        "url"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 10
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "path": {
                        "type": "string"
                      },
                      "scale": {
                        "type": "number",
                        "minimum": 0.1,
                        "maximum": 2
                      },
                      "displayHeaderFooter": {
                        "type": "boolean"
                      },
                      "headerTemplate": {
                        "type": "string"
                      },
                      "footerTemplate": {
                        "type": "string"
                      },
                      "printBackground": {
                        "type": "boolean"
                      },
                      "landscape": {
                        "type": "boolean"
                      },
                      "pageRanges": {
                        "type": "string"
                      },
                      "format": {
                        "type": "string",
                        "enum": [
                          "Letter",
                          "Legal",
                          "Tabloid",
                          "Ledger",
                          "A0",
                          "A1",
                          "A2",
                          "A3",
                          "A4",
                          "A5",
                          "A6"
                        ]
                      },
                      "width": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                      },
                      "height": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d+)?(mm|cm|in|px)$"
                      },
                      "margin": {
                        "type": "object",
                        "properties": {
                          "top": {
                            "type": "string"
                          },
                          "right": {
                            "type": "string"
                          },
                          "bottom": {
                            "type": "string"
                          },
                          "left": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      },
                      "preferCSSPageSize": {
                        "type": "boolean"
                      },
                      "omitBackground": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": [
                  "urls"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "jobIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "message",
                    "jobIds",
                    "errors"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}